Remove .net 4.5 project and move INetworkSocket implementation to .net standard project
